% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiNetEnrich.R
\name{MultiNetEnrich}
\alias{MultiNetEnrich}
\title{Perform functional network enrichment for multiple data sets}
\usage{
MultiNetEnrich(
  X,
  database = NULL,
  nperm = 1000,
  maxitr = 1000,
  scale = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{X}{list of the \eqn{K} observed \eqn{n_k} by \eqn{p} data matrices}

\item{database}{gene set database(s) to use from MSigDb. Provided as one or more strings containing the URL(s), from \url{https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.5.1/}.
If not specified, Wikipathways, PID and Hallmark are used.}

\item{nperm}{the number of random permuations to use. Default \eqn{1000}.}

\item{maxitr}{maximum number of iterations. Default \eqn{1000}.}

\item{scale}{should variables be scaled? Default \code{TRUE}}

\item{verbose}{logical indicator of printing information at each iteration}
}
\value{
a fitted \code{MultiNetEnrich} object
}
\description{
This function performs functional network enrichment for multiple data sets
}
